backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
requests
httpx
shopifyapi
qdrant-client
langchain
langgraph
google-generativeai
langchain-google-genai
pydantic
asyncio
stream-chat
python-multipart
jose
passlib
bcrypt
